Fix slasti bookmarklets
The bookmarklet must use encodeURIComponent(). Otherwise URLs containing
ampersand (&) chars breaks, because everything after the ampersand is
treated as additional parameter for "new", instead of part of the href:

Test link to bookmark:
  http://x.test/cgi?x=1&y=2
The bookmarklet generates this "add a new bookmark" URL for slasti:
  http://slasti.url/user/new?href=http://x.test/cgi?x=1&y=2&title=

In this URL, it's impossible to say if the &y=2 is part of the href arg
or a separate argument to /user/new. Indeed, urlparse.parse_qs() parses
this to:
{ "href": ["http://x.test/cgi?x=1"],
  "y": ["2"],
  # title empty, so it's not included in the dict
}

Thus the "&y=2" is swallowed and the URL that should be bookmarked is
corrupted. Using encodeURIComponent() solves this problem by quoting the
special characters. This quoting is automatically removed by
urlparse.parse_qs().

Next, urllib.quote_plus isn't the right function to use as it garbles
both titles and URLs.
cgi.escape(s, True) could work, but doesn't escape single quotes. Thus
you'd have to ensure that no one ever uses value='...' within the
templates. It's better to play safe and use a custom function instead.

With these changes, bookmarking the following torture-test works:
https://www.google.com/search?q=xyz"'> <span> JÖhö </span>&v=7\