Add bookmark search
Pretty straight forward. The string matching is done in
TagMark.contains() and is case-insensitive.

The only tricky part of the patch is search_back(). Since not every
bookmark matches, it has some freedom on which mark to return. It could
stop immediately at the 25th (PAGESZ'th) matching mark, or it could
return any non-matching mark further back to just after the 26th
(PAGESZ+1'th) mark.

The current implementation returns the earliest possible mark (i.e. the
PAGESZ+1 approach), as that's needed to avoid showing the "<<" button on
the bottom of the search page when no earlier matching bookmarks are
actually available.