Add SQLlite based data storage backend
Add SQLite support. Introduce a new class SlastiDB and matching DBMark
and DBTag classes. The database file is stored as bookmarks.db in the
user's bookmarks folder.

Swap out ctx.base to make use of the DB backend.

Also introduce a main() function in tagbase.py which can be used to
convert from file-backend to database-backend.