Parse loc: data from notes and show it on a map
Add a new tool to the top toolbar called "parse loc". It searches within
bookmark notes for location data in the form "loc:48.210247,16.373488".
Then it plots all the points of the current user on a map.

This is rough-and-ready code. It could use improvement:
 * Find a better way to plot the map. Currently, we use
   copypastemap.com. It sucks, but has an easy to use interface.
 * Allow adding a further search string to filter the loc: results.
 * Implement about a useful UI to enter said filter string and to
   display the results in-line.

Along the way, some unrelated changes were made:
 * Modify style.css, since the UI suddenly became messed up?!
   Maybe that was triggered by the move to HTML5 doctype earlier.
 * Add proper classes the generated HTML for bookmarks. This enables
   easy parsing of the bookmark data in javascript without needing extra
   json-export from the server code.