Switch from copypastemap to OpenLayers
Instead of displaying the loc: data externally on copypastemap.com,
display them on-page using the OpenLayers JavaScript maps library.

Map provider choice:
* OSM is free and easily accessible, but the maps are horrible
* Google looks good, but requires all of the proprietary Google Maps
  JavaScript library to be loaded
* Microsoft Bing Maps looks OK and can be accessed directly without
  further JS code. Meta-information is loaded and executed via JSON-P,
  unfortunately. It is harmless and verifyable code, however.

Thus, we use Bing Maps, but it's relatively easy to switch back to OSM
if complications arise.

Maybe we should reflect the initial JSON data load via the server, then
we could eliminate the need to execute code supplied by Microsoft.

We use OpenLayers' API key for now, we have to change this if we
actually publish this code!