Fix html entity handling in Markdown parser
The input to the Markdown parser was taken to be the innerHTML of the
notes element. This meant that characters such as & and <> were
encoded as HTML entities, and Markdown double-escaped those, so "&amp;"
showed up in user-visible output.

Use the raw, unescaped data as input for Markdown instead, but ensure
that doesn't allow injection of arbitrary HTML by including an HTML
sanitizer (from Google's caja project).