Make static files serveable by Apache
Serving static files directly from the slasti root URL means that it's
not practical to off-load serving those files to Apache. They have to be
in a separate sub-URL (sub-directory) instead.

Make that so. Currently, the sub-URL is hardcoded to 'static_files/',
that could be made configurable, but the benefits are probably thin.