Cache TF-IDF similarity data in DB
Calculating the TF-IDF similarity matrices for all documents in the
database takes a fair amount of time (~1 second). This additional delay
is fairly noticeable.

Create a caching layer in the user database and use it to cache the
intermediate results to speed up later calculations.

In the cache-hit case we're down to 10ms for similarity calculations.