Blacklist certain hosts in similarity search
The similarity search always returns a ton of Wikipedia articles as
"similar match" for any Wikipedia bookmark. That's not very productive.

One part of this is easily solved by stopwording "wikipedia" in the
title, but the host matcher also detects this, and is harder to handle
via stopwords. Consequently, we just drop the host if it is in the
user's "ignore_hosts_for_search" config option.