Show incoming slasti-internal links on the target
If one bookmark links to another, it's trivially obvious on the source
bookmark (and we have made it more obvious in the past by adding a white
star to all internal links).

Up until now, incoming links were invisible on the target bookmark.
Change this: track internal links in a separate DB table, and show
incoming links on the target bookmark.

This should give slasti more of a "web of knowledge" quality.