Persist unsaved notes when user navigates away
When the user typed in notes and navigated away from the page, the notes
were lost (ACE editor did nothing to save them). This happened in a
couple of scenarios:

* User hits the back button to return to the origin page
* User tries to save without tags and gets an error message

It was very frustrating if that happened after typing in multi-page
notes.

To alleviate this problem, save notes in LocalStorage and fetch them
back if no server-provided notes exist.